Merge request from develop into release

See merge request s08-final/S08P31C101!692

+++ backend/product-service/src/main/java/com/example/productservice/dto/request/ReservationRequestDto.java
@@ -16,9 +16,9 @@ public class ReservationRequestDto {
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd", timezone = "Asia/Seoul")
     private LocalDate endDate;
 
-    private Float lng;
+    private Double lng;
 
-    private Float lat;
+    private Double lat;
 
     private String location;
